Validation on purchase (used_balance + value)
There was an error when a purchase was being made.
The used_balance + the value being used was not being compared to the
actual wallet limit, in that case purchase didnt care for the limit
put in the wallet when making more than one purchase.

This commit attempts to fix this issue by adding one more condition on
`CreditCardService#purchase_allowed?` method